requirements.txt
confluent-kafka==2.6.0
python-dotenv==1.0.1
mysql-connector-python==9.1.0
numpy==2.1.2
